requirements.txt
altair==5.5.0
fastapi==0.115.6
flet==0.25.2
geopy==2.4.1
loguru==0.7.3
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
passlib==1.7.4
Pillow==11.0.0
pydantic==2.10.4
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
streamlit==1.41.1
torch==2.5.1
ultralytics==8.3.55
uvicorn==0.34.0
absl-py==2.1.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.6
filelock==3.16.1
flet==0.25.2
fonttools==4.55.3
fsspec==2024.12.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
loguru==0.7.3
mako==1.3.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.19.1
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.4
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.18.0
pyjwt==2.10.1
pymysql==1.1.1
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
repath==0.9.0
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.3
streamlit==1.41.1
streamlit-elements==0.1.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
triton==3.1.0
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
werkzeug==3.1.3
